add doorlock.cpp
